refactor: Split large files for maintainability
launcher_tui.py split:
- Created launcher_tui/ package
- Extracted DialogBackend to launcher_tui/backend.py (185 lines)
- Main launcher stays in launcher_tui/main.py (2,811 lines)
- Added __main__.py for `python -m launcher_tui` support

main_web.py refactoring:
- Extracted HTML templates to src/web/templates/
  - login.html (56 lines)
  - main.html (1,523 lines)
- Reduced main_web.py from 2,901 to 1,314 lines
- Using render_template() instead of render_template_string()
- Flask app now properly configured with template_folder

File size changes:
- launcher_tui.py: 2,982 → split into package (net same)
- main_web.py: 2,901 → 1,314 lines (-1,587 lines, -55%)

All syntax verified. Backwards compatible imports maintained.
