+++ src/launcher_tui/__main__.py
@@ -1,0 +1,8 @@
+"""
+Allow running the package directly with: python -m launcher_tui
+"""
+
+from . import main
+
+if __name__ == "__main__":
+    main()
