+++ src/main_web.py
@@ -34,12 +34,12 @@ _shutdown_flag = False
 sys.path.insert(0, str(Path(__file__).parent))
 
 try:
-    from flask import Flask, render_template_string, jsonify, request, redirect, url_for, session
+    from flask import Flask, render_template, render_template_string, jsonify, request, redirect, url_for, session
 except ImportError:
     print("Flask not installed. Installing...")
     subprocess.run([sys.executable, '-m', 'pip', 'install', '--break-system-packages', 'flask'],
                    capture_output=True, timeout=120)
-    from flask import Flask, render_template_string, jsonify, request, redirect, url_for, session
+    from flask import Flask, render_template, render_template_string, jsonify, request, redirect, url_for, session
 
 # Import centralized service checker
 try:
@@ -57,7 +57,9 @@ except ImportError:
     MeshtasticConnectionManager = None
     _meshtastic_mgr = None
 
-app = Flask(__name__)
+# Configure Flask with external templates
+_template_folder = Path(__file__).parent / 'web' / 'templates'
+app = Flask(__name__, template_folder=str(_template_folder))
 app.secret_key = secrets.token_hex(32)
 
 # Register modular blueprints (new architecture)
@@ -290,8 +292,8 @@ def login():
         if secrets.compare_digest(user_password, stored_password):
             session['authenticated'] = True
             return redirect(url_for('index'))
-        return render_template_string(LOGIN_TEMPLATE, error="Invalid password")
-    return render_template_string(LOGIN_TEMPLATE, error=None)
+        return render_template('login.html', error="Invalid password")
+    return render_template('login.html', error=None)
 
 
 @app.route('/logout')
@@ -1016,1595 +1018,6 @@ def send_mesh_message(text, destination=None):
 
 
 
-# ============================================================================
-# HTML Templates
-# ============================================================================
-
-LOGIN_TEMPLATE = '''
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Meshtasticd Manager - Login</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <style>
-        * { box-sizing: border-box; margin: 0; padding: 0; }
-        body {
-            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-            background: linear-gradient(135deg, #1a1a2e 0%, #16213e 100%);
-            min-height: 100vh;
-            display: flex;
-            align-items: center;
-            justify-content: center;
-        }
-        .login-box {
-            background: #fff;
-            padding: 40px;
-            border-radius: 10px;
-            box-shadow: 0 10px 40px rgba(0,0,0,0.3);
-            width: 100%;
-            max-width: 400px;
-        }
-        h1 { color: #333; margin-bottom: 30px; text-align: center; }
-        input[type="password"] {
-            width: 100%;
-            padding: 15px;
-            border: 2px solid #ddd;
-            border-radius: 5px;
-            font-size: 16px;
-            margin-bottom: 20px;
-        }
-        button {
-            width: 100%;
-            padding: 15px;
-            background: #4CAF50;
-            color: white;
-            border: none;
-            border-radius: 5px;
-            font-size: 16px;
-            cursor: pointer;
-        }
-        button:hover { background: #45a049; }
-        .error { color: #f44336; margin-bottom: 20px; text-align: center; }
-    </style>
-</head>
-<body>
-    <div class="login-box">
-        <h1><svg viewBox="0 0 128 128" width="40" height="40" style="vertical-align: middle; margin-right: 10px;"><defs><linearGradient id="bgGrad" x1="0%" y1="0%" x2="100%" y2="100%"><stop offset="0%" style="stop-color:#1a1a2e"/><stop offset="100%" style="stop-color:#16213e"/></linearGradient><linearGradient id="handGrad" x1="0%" y1="0%" x2="100%" y2="100%"><stop offset="0%" style="stop-color:#4CAF50"/><stop offset="100%" style="stop-color:#2E7D32"/></linearGradient></defs><circle cx="64" cy="64" r="60" fill="url(#bgGrad)"/><circle cx="64" cy="64" r="56" fill="none" stroke="#4CAF50" stroke-width="2" opacity="0.6"/><g fill="url(#handGrad)" stroke="#2E7D32" stroke-width="1.5"><path d="M64 92 C52 92 42 82 42 66 L42 56 C42 50 46 46 54 46 L74 46 C82 46 86 50 86 56 L86 66 C86 82 76 92 64 92 Z"/><path d="M42 60 C34 56 26 52 22 56 C18 60 18 68 22 72 C26 76 34 76 42 72"/><path d="M78 46 L82 28 C82 24 86 20 90 20 C94 20 98 24 98 28 L94 46"/></g></svg> MeshForge</h1>
-        {% if error %}<div class="error">{{ error }}</div>{% endif %}
-        <form method="POST">
-            <input type="password" name="password" placeholder="Password" autofocus>
-            <button type="submit">Login</button>
-        </form>
-    </div>
-</body>
-</html>
-'''
-
-MAIN_TEMPLATE = '''
-<!DOCTYPE html>
-<html>
-<head>
-    <title>MeshForge</title>
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <!-- Leaflet.js for maps -->
-    <link rel="stylesheet" href="https://unpkg.com/leaflet@1.9.4/dist/leaflet.css" crossorigin=""/>
-    <script src="https://unpkg.com/leaflet@1.9.4/dist/leaflet.js" crossorigin=""></script>
-    <style>
-        * { box-sizing: border-box; margin: 0; padding: 0; }
-        :root {
-            --bg-dark: #1a1a2e;
-            --bg-card: #16213e;
-            --bg-card-hover: #1e2a4a;
-            --text: #eee;
-            --text-muted: #888;
-            --accent: #4CAF50;
-            --accent-hover: #45a049;
-            --accent-glow: rgba(76, 175, 80, 0.3);
-            --danger: #f44336;
-            --warning: #ff9800;
-            --info: #2196F3;
-            --shadow-sm: 0 2px 4px rgba(0,0,0,0.2);
-            --shadow-md: 0 4px 12px rgba(0,0,0,0.3);
-            --shadow-lg: 0 8px 24px rgba(0,0,0,0.4);
-            --transition-fast: 0.15s ease;
-            --transition-normal: 0.25s ease;
-        }
-        body {
-            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-            background: linear-gradient(135deg, var(--bg-dark) 0%, #0f0f1a 100%);
-            color: var(--text);
-            min-height: 100vh;
-        }
-        .header {
-            background: linear-gradient(90deg, var(--bg-card) 0%, #1a2744 100%);
-            padding: 15px 20px;
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-            border-bottom: 2px solid var(--accent);
-            box-shadow: var(--shadow-md);
-            position: sticky;
-            top: 0;
-            z-index: 100;
-        }
-        .header h1 {
-            font-size: 1.5rem;
-            display: flex;
-            align-items: center;
-            gap: 10px;
-        }
-        .header h1 .logo {
-            font-size: 1.8rem;
-        }
-        .header .version { color: var(--text-muted); font-size: 0.85rem; }
-        .nav {
-            display: flex;
-            gap: 10px;
-        }
-        .nav a, .nav button {
-            padding: 8px 16px;
-            background: transparent;
-            color: var(--text);
-            border: 1px solid #444;
-            border-radius: 5px;
-            text-decoration: none;
-            cursor: pointer;
-            font-size: 14px;
-            transition: all var(--transition-fast);
-        }
-        .nav a:hover, .nav button:hover {
-            background: #333;
-            border-color: var(--accent);
-            transform: translateY(-1px);
-        }
-        .nav a.active {
-            background: var(--accent);
-            border-color: var(--accent);
-            box-shadow: 0 0 10px var(--accent-glow);
-        }
-
-        .container {
-            max-width: 1400px;
-            margin: 0 auto;
-            padding: 20px;
-        }
-
-        .grid {
-            display: grid;
-            grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-            gap: 20px;
-            margin-bottom: 20px;
-        }
-
-        .card {
-            background: var(--bg-card);
-            border-radius: 12px;
-            padding: 20px;
-            border: 1px solid #333;
-            box-shadow: var(--shadow-sm);
-            transition: all var(--transition-normal);
-        }
-        .card:hover {
-            border-color: #444;
-            box-shadow: var(--shadow-md);
-            transform: translateY(-2px);
-        }
-        .card h2 {
-            font-size: 0.85rem;
-            margin-bottom: 15px;
-            color: var(--text-muted);
-            text-transform: uppercase;
-            letter-spacing: 1.5px;
-            font-weight: 600;
-        }
-        .card .value {
-            font-size: 2rem;
-            font-weight: bold;
-            transition: color var(--transition-fast);
-        }
-        .card .value.success { color: var(--accent); text-shadow: 0 0 10px var(--accent-glow); }
-        .card .value.error { color: var(--danger); }
-        .card .value.warning { color: var(--warning); }
-
-        .progress-bar {
-            background: #222;
-            border-radius: 6px;
-            height: 8px;
-            margin-top: 10px;
-            overflow: hidden;
-            box-shadow: inset 0 1px 3px rgba(0,0,0,0.3);
-        }
-        .progress-bar .fill {
-            height: 100%;
-            background: linear-gradient(90deg, var(--accent) 0%, #66BB6A 100%);
-            transition: width 0.5s ease-out;
-            border-radius: 6px;
-            box-shadow: 0 0 8px var(--accent-glow);
-        }
-        .progress-bar .fill.warning { background: linear-gradient(90deg, var(--warning) 0%, #FFB74D 100%); }
-        .progress-bar .fill.danger { background: linear-gradient(90deg, var(--danger) 0%, #EF5350 100%); }
-
-        .stat-row {
-            display: flex;
-            justify-content: space-between;
-            padding: 12px 0;
-            border-bottom: 1px solid #333;
-            transition: background var(--transition-fast);
-        }
-        .stat-row:hover { background: rgba(255,255,255,0.02); }
-        .stat-row:last-child { border-bottom: none; }
-        .stat-label { color: var(--text-muted); }
-
-        .btn {
-            padding: 10px 20px;
-            border: none;
-            border-radius: 6px;
-            cursor: pointer;
-            font-size: 14px;
-            font-weight: 500;
-            margin: 5px;
-            transition: all var(--transition-fast);
-            text-transform: uppercase;
-            letter-spacing: 0.5px;
-        }
-        .btn-success {
-            background: linear-gradient(135deg, var(--accent) 0%, #45a049 100%);
-            color: white;
-            box-shadow: 0 2px 8px var(--accent-glow);
-        }
-        .btn-success:hover {
-            transform: translateY(-1px);
-            box-shadow: 0 4px 12px var(--accent-glow);
-        }
-        .btn-danger {
-            background: linear-gradient(135deg, var(--danger) 0%, #d32f2f 100%);
-            color: white;
-        }
-        .btn-danger:hover { transform: translateY(-1px); }
-        .btn-warning {
-            background: linear-gradient(135deg, var(--warning) 0%, #f57c00 100%);
-            color: black;
-        }
-        .btn-warning:hover { transform: translateY(-1px); }
-        .btn:disabled { opacity: 0.5; cursor: not-allowed; transform: none !important; }
-
-        .log-box {
-            background: #0d0d0d;
-            border-radius: 8px;
-            padding: 15px;
-            font-family: 'SF Mono', 'Fira Code', 'Consolas', monospace;
-            font-size: 12px;
-            line-height: 1.5;
-            max-height: 300px;
-            overflow-y: auto;
-            white-space: pre-wrap;
-            word-break: break-all;
-            border: 1px solid #222;
-            box-shadow: inset 0 2px 4px rgba(0,0,0,0.3);
-        }
-        .log-box::-webkit-scrollbar { width: 8px; }
-        .log-box::-webkit-scrollbar-track { background: #111; }
-        .log-box::-webkit-scrollbar-thumb { background: #444; border-radius: 4px; }
-        .log-box::-webkit-scrollbar-thumb:hover { background: #555; }
-
-        .config-list {
-            list-style: none;
-        }
-        .config-item {
-            display: flex;
-            justify-content: space-between;
-            align-items: center;
-            padding: 12px 15px;
-            background: #1a1a2a;
-            margin: 8px 0;
-            border-radius: 8px;
-            border: 1px solid #333;
-            transition: all var(--transition-fast);
-        }
-        .config-item:hover {
-            background: #222;
-            border-color: #444;
-        }
-        .config-item .name { font-family: monospace; font-size: 0.95rem; }
-
-        .hardware-item {
-            display: flex;
-            align-items: center;
-            padding: 12px 15px;
-            background: #1a1a2a;
-            margin: 8px 0;
-            border-radius: 8px;
-            border: 1px solid #333;
-            transition: all var(--transition-fast);
-        }
-        .hardware-item:hover {
-            background: #222;
-            border-color: #444;
-        }
-        .hardware-item .badge {
-            padding: 4px 10px;
-            border-radius: 4px;
-            font-size: 11px;
-            font-weight: 600;
-            margin-right: 12px;
-            text-transform: uppercase;
-            letter-spacing: 0.5px;
-        }
-        .hardware-item .badge.active { background: var(--accent); color: white; }
-        .hardware-item .badge.spi { background: var(--info); color: white; }
-        .hardware-item .badge.i2c { background: #9C27B0; color: white; }
-        .hardware-item .badge.info { background: #607D8B; color: white; }
-
-        .tabs {
-            display: flex;
-            gap: 5px;
-            border-bottom: 2px solid #333;
-            margin-bottom: 25px;
-            padding-bottom: 0;
-            overflow-x: auto;
-            scrollbar-width: none;
-        }
-        .tabs::-webkit-scrollbar { display: none; }
-        .tab {
-            padding: 12px 20px;
-            cursor: pointer;
-            border-bottom: 3px solid transparent;
-            margin-bottom: -2px;
-            color: var(--text-muted);
-            font-weight: 500;
-            font-size: 0.9rem;
-            transition: all var(--transition-fast);
-            white-space: nowrap;
-            border-radius: 8px 8px 0 0;
-        }
-        .tab:hover {
-            color: var(--text);
-            background: rgba(255,255,255,0.03);
-        }
-        .tab.active {
-            color: var(--accent);
-            border-bottom-color: var(--accent);
-            background: rgba(76,175,80,0.1);
-        }
-        .tab-content { display: none; animation: fadeIn 0.3s ease; }
-        .tab-content.active { display: block; }
-        @keyframes fadeIn {
-            from { opacity: 0; transform: translateY(10px); }
-            to { opacity: 1; transform: translateY(0); }
-        }
-
-        .radio-info {
-            display: grid;
-            grid-template-columns: repeat(2, 1fr);
-            gap: 12px;
-        }
-        .radio-info .item {
-            padding: 15px;
-            background: #1a1a2a;
-            border-radius: 8px;
-            border: 1px solid #333;
-            transition: all var(--transition-fast);
-        }
-        .radio-info .item:hover {
-            background: #222;
-            border-color: #444;
-        }
-        .radio-info .label {
-            color: var(--text-muted);
-            font-size: 11px;
-            text-transform: uppercase;
-            letter-spacing: 1px;
-            margin-bottom: 8px;
-        }
-        .radio-info .value {
-            font-size: 1.1rem;
-            font-weight: 500;
-        }
-
-        /* Input styling */
-        input, textarea, select {
-            transition: border-color var(--transition-fast), box-shadow var(--transition-fast);
-        }
-        input:focus, textarea:focus, select:focus {
-            outline: none;
-            border-color: var(--accent) !important;
-            box-shadow: 0 0 0 3px var(--accent-glow);
-        }
-
-        /* Link styling */
-        a {
-            color: var(--accent);
-            text-decoration: none;
-            transition: color var(--transition-fast);
-        }
-        a:hover {
-            color: #66BB6A;
-            text-decoration: underline;
-        }
-
-        /* Table styling */
-        table {
-            border-collapse: collapse;
-            width: 100%;
-        }
-        th {
-            text-align: left;
-            font-weight: 600;
-            color: var(--text-muted);
-            text-transform: uppercase;
-            font-size: 0.8rem;
-            letter-spacing: 1px;
-        }
-        tr {
-            transition: background var(--transition-fast);
-        }
-        tr:hover {
-            background: rgba(255,255,255,0.02);
-        }
-
-        /* Footer branding */
-        .footer {
-            text-align: center;
-            padding: 20px;
-            color: var(--text-muted);
-            font-size: 0.85rem;
-            margin-top: 40px;
-            border-top: 1px solid #333;
-        }
-
-        @media (max-width: 768px) {
-            .grid { grid-template-columns: 1fr; }
-            .header { flex-direction: column; gap: 10px; }
-            .nav { flex-wrap: wrap; justify-content: center; }
-            .tabs { gap: 2px; }
-            .tab { padding: 10px 15px; font-size: 0.85rem; }
-            .radio-info { grid-template-columns: 1fr; }
-        }
-    </style>
-</head>
-<body>
-    <div class="header">
-        <div>
-            <h1><span class="logo"><svg viewBox="0 0 128 128" width="32" height="32" style="vertical-align: middle; margin-right: 8px;"><defs><linearGradient id="bgGrad" x1="0%" y1="0%" x2="100%" y2="100%"><stop offset="0%" style="stop-color:#1a1a2e"/><stop offset="100%" style="stop-color:#16213e"/></linearGradient><linearGradient id="handGrad" x1="0%" y1="0%" x2="100%" y2="100%"><stop offset="0%" style="stop-color:#4CAF50"/><stop offset="100%" style="stop-color:#2E7D32"/></linearGradient></defs><circle cx="64" cy="64" r="60" fill="url(#bgGrad)"/><circle cx="64" cy="64" r="56" fill="none" stroke="#4CAF50" stroke-width="2" opacity="0.6"/><circle cx="24" cy="32" r="4" fill="#2196F3" opacity="0.5"/><circle cx="104" cy="32" r="4" fill="#2196F3" opacity="0.5"/><circle cx="24" cy="96" r="4" fill="#2196F3" opacity="0.5"/><circle cx="104" cy="96" r="4" fill="#2196F3" opacity="0.5"/><line x1="24" y1="32" x2="64" y2="64" stroke="#2196F3" stroke-width="1" opacity="0.3"/><line x1="104" y1="32" x2="64" y2="64" stroke="#2196F3" stroke-width="1" opacity="0.3"/><line x1="24" y1="96" x2="64" y2="64" stroke="#2196F3" stroke-width="1" opacity="0.3"/><line x1="104" y1="96" x2="64" y2="64" stroke="#2196F3" stroke-width="1" opacity="0.3"/><g fill="url(#handGrad)" stroke="#2E7D32" stroke-width="1.5"><path d="M64 92 C52 92 42 82 42 66 L42 56 C42 50 46 46 54 46 L74 46 C82 46 86 50 86 56 L86 66 C86 82 76 92 64 92 Z"/><path d="M42 60 C34 56 26 52 22 56 C18 60 18 68 22 72 C26 76 34 76 42 72"/><path d="M78 46 L82 28 C82 24 86 20 90 20 C94 20 98 24 98 28 L94 46"/><path d="M54 46 L54 40 C54 36 58 34 62 36 L62 46" opacity="0.8"/><path d="M62 46 L62 36 C62 32 66 30 70 32 L70 46" opacity="0.8"/><path d="M70 46 L70 38 C70 34 74 32 78 34 L78 46" opacity="0.8"/><path d="M52 92 L52 104 L76 104 L76 92"/></g></svg></span> MeshForge</h1>
-            <span class="version">v{{ version }} | Mesh Network Operations Center</span>
-        </div>
-        <div class="nav">
-            {% if auth_enabled %}
-            <a href="/logout">Logout</a>
-            {% endif %}
-        </div>
-    </div>
-
-    <div class="container">
-        <div class="tabs">
-            <div class="tab active" data-tab="dashboard">Dashboard</div>
-            <div class="tab" data-tab="gateway">Gateway</div>
-            <div class="tab" data-tab="map">Map</div>
-            <div class="tab" data-tab="nodes">Nodes</div>
-            <div class="tab" data-tab="messages">Messages</div>
-            <div class="tab" data-tab="service">Service</div>
-            <div class="tab" data-tab="config">Config</div>
-            <div class="tab" data-tab="hardware">Hardware</div>
-            <div class="tab" data-tab="radio">Radio</div>
-            <div class="tab" data-tab="updates">Updates</div>
-            <div class="tab" data-tab="system">System</div>
-        </div>
-
-        <!-- Dashboard Tab -->
-        <div id="dashboard" class="tab-content active">
-            <div class="grid">
-                <div class="card">
-                    <h2>Service Status</h2>
-                    <div class="value" id="service-status">Loading...</div>
-                </div>
-                <div class="card">
-                    <h2>CPU Usage</h2>
-                    <div class="value" id="cpu-value">--</div>
-                    <div class="progress-bar"><div class="fill" id="cpu-bar" style="width: 0%"></div></div>
-                </div>
-                <div class="card">
-                    <h2>Memory</h2>
-                    <div class="value" id="mem-value">--</div>
-                    <div class="progress-bar"><div class="fill" id="mem-bar" style="width: 0%"></div></div>
-                </div>
-                <div class="card">
-                    <h2>Disk Usage</h2>
-                    <div class="value" id="disk-value">--</div>
-                    <div class="progress-bar"><div class="fill" id="disk-bar" style="width: 0%"></div></div>
-                </div>
-                <div class="card">
-                    <h2>Temperature</h2>
-                    <div class="value" id="temp-value">--</div>
-                    <div class="progress-bar"><div class="fill" id="temp-bar" style="width: 0%"></div></div>
-                </div>
-                <div class="card">
-                    <h2>Uptime</h2>
-                    <div class="value" id="uptime-value">--</div>
-                </div>
-            </div>
-
-            <div class="card">
-                <h2>Recent Logs</h2>
-                <div class="log-box" id="logs">Loading...</div>
-            </div>
-        </div>
-
-        <!-- Gateway Tab -->
-        <div id="gateway" class="tab-content">
-            <div class="grid">
-                <div class="card" style="grid-column: span 2;">
-                    <div style="display: flex; justify-content: space-between; align-items: center; margin-bottom: 15px;">
-                        <h2 style="margin: 0;">Meshtastic ↔ RNS Gateway Bridge</h2>
-                        <div>
-                            <button class="btn btn-success" onclick="startGateway()" id="gateway-start-btn">Start</button>
-                            <button class="btn btn-danger" onclick="stopGateway()" id="gateway-stop-btn">Stop</button>
-                            <button class="btn" onclick="refreshGateway()">Refresh</button>
-                        </div>
-                    </div>
-                    <div class="grid" style="grid-template-columns: repeat(4, 1fr); gap: 15px;">
-                        <div style="text-align: center; padding: 15px; background: var(--card-bg); border-radius: 8px;">
-                            <div style="font-size: 0.9em; color: var(--text-muted);">Gateway Status</div>
-                            <div id="gateway-status" style="font-size: 1.5em; font-weight: bold; margin-top: 5px;">--</div>
-                        </div>
-                        <div style="text-align: center; padding: 15px; background: var(--card-bg); border-radius: 8px;">
-                            <div style="font-size: 0.9em; color: var(--text-muted);">meshtasticd</div>
-                            <div id="gateway-mesh-status" style="font-size: 1.5em; font-weight: bold; margin-top: 5px;">--</div>
-                        </div>
-                        <div style="text-align: center; padding: 15px; background: var(--card-bg); border-radius: 8px;">
-                            <div style="font-size: 0.9em; color: var(--text-muted);">rnsd</div>
-                            <div id="gateway-rns-status" style="font-size: 1.5em; font-weight: bold; margin-top: 5px;">--</div>
-                        </div>
-                        <div style="text-align: center; padding: 15px; background: var(--card-bg); border-radius: 8px;">
-                            <div style="font-size: 0.9em; color: var(--text-muted);">Tracked Nodes</div>
-                            <div id="gateway-nodes" style="font-size: 1.5em; font-weight: bold; margin-top: 5px;">--</div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <div class="grid" style="margin-top: 20px;">
-                <div class="card">
-                    <h2>Bridge Statistics</h2>
-                    <div id="gateway-stats" style="display: grid; grid-template-columns: 1fr 1fr; gap: 10px;">
-                        <div class="item"><div class="label">Mesh → RNS</div><div id="stat-mesh-to-rns">0</div></div>
-                        <div class="item"><div class="label">RNS → Mesh</div><div id="stat-rns-to-mesh">0</div></div>
-                        <div class="item"><div class="label">Meshtastic Nodes</div><div id="stat-mesh-nodes">0</div></div>
-                        <div class="item"><div class="label">RNS Nodes</div><div id="stat-rns-nodes">0</div></div>
-                        <div class="item"><div class="label">Errors</div><div id="stat-errors">0</div></div>
-                        <div class="item"><div class="label">Last Activity</div><div id="stat-last-activity">--</div></div>
-                    </div>
-                </div>
-                <div class="card">
-                    <h2>Diagnostics</h2>
-                    <button class="btn btn-success" onclick="runGatewayDiagnostic()" style="margin-bottom: 15px;">Run Diagnostic</button>
-                    <button class="btn" onclick="testGatewayConnections()" style="margin-bottom: 15px;">Test Connections</button>
-                    <div id="gateway-diagnostic" class="log-box" style="max-height: 200px; overflow-y: auto;">Click "Run Diagnostic" to check gateway prerequisites</div>
-                </div>
-            </div>
-
-            <div class="card" style="margin-top: 20px;">
-                <h2>Tracked Nodes</h2>
-                <button class="btn btn-success" onclick="refreshGatewayNodes()" style="margin-bottom: 15px;">Refresh</button>
-                <div id="gateway-node-list">
-                    <table style="width: 100%; border-collapse: collapse;">
-                        <tr style="border-bottom: 1px solid #444;">
-                            <th style="padding: 8px; text-align: left;">Network</th>
-                            <th style="padding: 8px; text-align: left;">ID</th>
-                            <th style="padding: 8px; text-align: left;">Name</th>
-                            <th style="padding: 8px; text-align: left;">Last Seen</th>
-                        </tr>
-                    </table>
-                    <div style="color: var(--text-muted); padding: 10px;">No nodes tracked yet</div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Map Tab -->
-        <div id="map" class="tab-content">
-            <div class="card" style="padding: 10px;">
-                <div style="display: flex; justify-content: space-between; align-items: center; margin-bottom: 10px;">
-                    <h2 style="margin: 0;">Mesh Network Map</h2>
-                    <div>
-                        <span id="map-node-count" style="color: var(--text-muted); margin-right: 15px;">--</span>
-                        <button class="btn btn-success" onclick="refreshMap()">Refresh Map</button>
-                    </div>
-                </div>
-                <div id="mesh-map" style="height: 500px; border-radius: 8px; background: #1a1a2e;"></div>
-                <div id="map-legend" style="margin-top: 10px; padding: 10px; background: #222; border-radius: 5px;">
-                    <span style="margin-right: 20px;"><span style="display: inline-block; width: 12px; height: 12px; background: #4CAF50; border-radius: 50%; margin-right: 5px;"></span> My Node</span>
-                    <span style="margin-right: 20px;"><span style="display: inline-block; width: 12px; height: 12px; background: #2196F3; border-radius: 50%; margin-right: 5px;"></span> Online (< 1h)</span>
-                    <span style="margin-right: 20px;"><span style="display: inline-block; width: 12px; height: 12px; background: #ff9800; border-radius: 50%; margin-right: 5px;"></span> Stale (1-24h)</span>
-                    <span><span style="display: inline-block; width: 12px; height: 12px; background: #607D8B; border-radius: 50%; margin-right: 5px;"></span> Offline (> 24h)</span>
-                </div>
-            </div>
-            <div class="card" style="margin-top: 20px;">
-                <h2>Nodes with Position</h2>
-                <div id="map-node-list" style="max-height: 300px; overflow-y: auto;">Loading...</div>
-            </div>
-        </div>
-
-        <!-- Nodes Tab -->
-        <div id="nodes" class="tab-content">
-            <div class="card">
-                <h2>Mesh Nodes</h2>
-                <button class="btn btn-success" onclick="refreshNodes()" style="margin-bottom: 15px;">Refresh Nodes</button>
-                <div id="nodes-list">Loading...</div>
-                <div id="nodes-raw" style="margin-top: 15px; display: none;">
-                    <h3>Raw Output</h3>
-                    <div class="log-box" id="nodes-raw-content" style="max-height: 200px;"></div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Messages Tab -->
-        <div id="messages" class="tab-content">
-            <div class="card">
-                <h2>Send Message</h2>
-                <div style="margin-bottom: 15px;">
-                    <label style="display: block; margin-bottom: 5px; color: var(--text-muted);">Message Text:</label>
-                    <textarea id="message-text" rows="3" style="width: 100%; padding: 10px; background: #222; border: 1px solid #444; border-radius: 5px; color: var(--text); font-family: inherit;" placeholder="Type your message here..."></textarea>
-                </div>
-                <div style="margin-bottom: 15px;">
-                    <label style="display: block; margin-bottom: 5px; color: var(--text-muted);">Destination (optional, leave blank for broadcast):</label>
-                    <input type="text" id="message-dest" style="width: 100%; padding: 10px; background: #222; border: 1px solid #444; border-radius: 5px; color: var(--text);" placeholder="!abcd1234 or leave blank">
-                </div>
-                <button class="btn btn-success" onclick="sendMessage()">Send Message</button>
-                <div id="message-status" style="margin-top: 15px;"></div>
-            </div>
-            <div class="card" style="margin-top: 20px;">
-                <h2>Recent Messages</h2>
-                <p style="color: var(--text-muted);">Message history coming soon. Check the logs in the Service tab for now.</p>
-            </div>
-        </div>
-
-        <!-- Service Tab -->
-        <div id="service" class="tab-content">
-            <div class="card">
-                <h2>Service Control</h2>
-                <p style="margin-bottom: 20px; color: var(--text-muted);">
-                    Control the meshtasticd service
-                </p>
-                <button class="btn btn-success" onclick="serviceAction('start')">Start</button>
-                <button class="btn btn-danger" onclick="serviceAction('stop')">Stop</button>
-                <button class="btn btn-warning" onclick="serviceAction('restart')">Restart</button>
-            </div>
-
-            <div class="card" style="margin-top: 20px;">
-                <h2>Service Logs</h2>
-                <button class="btn btn-success" onclick="refreshLogs()" style="margin-bottom: 15px;">Refresh Logs</button>
-                <div class="log-box" id="service-logs" style="max-height: 500px;">Loading...</div>
-            </div>
-        </div>
-
-        <!-- Config Tab -->
-        <div id="config" class="tab-content">
-            <div class="grid">
-                <div class="card">
-                    <h2>Active Configurations</h2>
-                    <ul class="config-list" id="active-configs">Loading...</ul>
-                </div>
-                <div class="card">
-                    <h2>Available Configurations</h2>
-                    <ul class="config-list" id="available-configs">Loading...</ul>
-                </div>
-            </div>
-        </div>
-
-        <!-- Hardware Tab -->
-        <div id="hardware" class="tab-content">
-            <div class="card">
-                <h2>Detected Hardware</h2>
-                <button class="btn btn-success" onclick="refreshHardware()" style="margin-bottom: 15px;">Refresh</button>
-                <div id="hardware-list">Loading...</div>
-            </div>
-        </div>
-
-        <!-- Radio Tab -->
-        <div id="radio" class="tab-content">
-            <div class="card">
-                <h2>Connected Radio</h2>
-                <button class="btn btn-success" onclick="refreshRadio(true)" style="margin-bottom: 15px;">Refresh</button>
-                <div class="radio-info" id="radio-info">Loading...</div>
-            </div>
-        </div>
-
-        <!-- Updates Tab -->
-        <div id="updates" class="tab-content">
-            <div class="card">
-                <h2>Component Versions</h2>
-                <p style="color: var(--text-muted); margin-bottom: 15px;">
-                    Check for updates to meshtasticd, CLI, and node firmware.
-                </p>
-                <button class="btn btn-success" onclick="refreshVersions()" style="margin-bottom: 15px;">Check for Updates</button>
-                <div id="version-list">Loading...</div>
-                <div id="version-status" style="margin-top: 15px; color: var(--text-muted);"></div>
-            </div>
-            <div class="card" style="margin-top: 20px;">
-                <h2>Update Instructions</h2>
-                <div style="color: var(--text-muted);">
-                    <h3 style="color: var(--text); margin: 15px 0 10px 0;">meshtasticd</h3>
-                    <pre style="background: #111; padding: 10px; border-radius: 5px; overflow-x: auto;">sudo apt update && sudo apt upgrade meshtasticd</pre>
-
-                    <h3 style="color: var(--text); margin: 15px 0 10px 0;">Meshtastic CLI</h3>
-                    <pre style="background: #111; padding: 10px; border-radius: 5px; overflow-x: auto;">pipx upgrade meshtastic</pre>
-
-                    <h3 style="color: var(--text); margin: 15px 0 10px 0;">Node Firmware</h3>
-                    <p>Use the Meshtastic Web Flasher or the meshtastic-flasher CLI tool.</p>
-                </div>
-            </div>
-        </div>
-
-        <!-- System Tab -->
-        <div id="system" class="tab-content">
-            <div class="card">
-                <h2>Top Processes</h2>
-                <button class="btn btn-success" onclick="refreshProcesses()" style="margin-bottom: 15px;">Refresh</button>
-                <div class="log-box" id="processes">Loading...</div>
-            </div>
-        </div>
-    </div>
-
-    <script>
-        // Security: HTML escape function to prevent XSS
-        function escapeHtml(text) {
-            if (text === null || text === undefined) return '';
-            const div = document.createElement('div');
-            div.textContent = String(text);
-            return div.innerHTML;
-        }
-
-        // Tab switching with persistence
-        function switchTab(tabName) {
-            document.querySelectorAll('.tab').forEach(t => t.classList.remove('active'));
-            document.querySelectorAll('.tab-content').forEach(c => c.classList.remove('active'));
-            const tab = document.querySelector(`.tab[data-tab="${tabName}"]`);
-            if (tab) {
-                tab.classList.add('active');
-                document.getElementById(tabName).classList.add('active');
-                localStorage.setItem('activeTab', tabName);
-            }
-        }
-
-        document.querySelectorAll('.tab').forEach(tab => {
-            tab.addEventListener('click', () => switchTab(tab.dataset.tab));
-        });
-
-        // Restore last active tab
-        const savedTab = localStorage.getItem('activeTab');
-        if (savedTab && document.getElementById(savedTab)) {
-            switchTab(savedTab);
-        }
-
-        // API calls
-        async function fetchStatus() {
-            try {
-                const resp = await fetch('/api/status');
-                const data = await resp.json();
-
-                // Service status
-                const statusEl = document.getElementById('service-status');
-                statusEl.textContent = data.service.status;
-                statusEl.className = 'value ' + (data.service.running ? 'success' : 'error');
-
-                // CPU
-                document.getElementById('cpu-value').textContent = data.system.cpu_percent + '%';
-                document.getElementById('cpu-bar').style.width = data.system.cpu_percent + '%';
-
-                // Memory
-                document.getElementById('mem-value').textContent =
-                    data.system.mem_used_mb + '/' + data.system.mem_total_mb + ' MB';
-                document.getElementById('mem-bar').style.width = data.system.mem_percent + '%';
-
-                // Disk
-                document.getElementById('disk-value').textContent =
-                    data.system.disk_used_gb + '/' + data.system.disk_total_gb + ' GB';
-                document.getElementById('disk-bar').style.width = data.system.disk_percent + '%';
-
-                // Temperature
-                if (data.system.temperature) {
-                    document.getElementById('temp-value').textContent = data.system.temperature + '°C';
-                    const tempPct = Math.min(data.system.temperature / 85 * 100, 100);
-                    document.getElementById('temp-bar').style.width = tempPct + '%';
-                }
-
-                // Uptime
-                document.getElementById('uptime-value').textContent = data.system.uptime;
-            } catch (e) {
-                console.error('Error fetching status:', e);
-            }
-        }
-
-        async function fetchLogs() {
-            try {
-                const resp = await fetch('/api/logs?lines=30');
-                const data = await resp.json();
-                document.getElementById('logs').textContent = data.logs;
-                document.getElementById('service-logs').textContent = data.logs;
-            } catch (e) {
-                console.error('Error fetching logs:', e);
-            }
-        }
-
-        async function refreshLogs() {
-            const resp = await fetch('/api/logs?lines=100');
-            const data = await resp.json();
-            document.getElementById('service-logs').textContent = data.logs;
-        }
-
-        async function fetchConfigs() {
-            try {
-                const resp = await fetch('/api/configs');
-                const data = await resp.json();
-
-                const activeEl = document.getElementById('active-configs');
-                const availEl = document.getElementById('available-configs');
-                let activeHtml = '';
-
-                // Ensure arrays exist
-                const activeConfigs = Array.isArray(data.active) ? data.active : [];
-                const availableConfigs = Array.isArray(data.available) ? data.available : [];
-
-                // Show main config if exists
-                if (data.main_config) {
-                    activeHtml += `<li class="config-item"><span class="name" style="color: var(--accent);">📄 ${escapeHtml(data.main_config)}</span></li>`;
-                }
-
-                // Show error if any
-                if (data.error) {
-                    activeHtml += `<li class="config-item" style="color: var(--warning);">${escapeHtml(data.error)}</li>`;
-                }
-
-                // Show active configs from config.d
-                if (activeConfigs.length > 0) {
-                    activeHtml += activeConfigs.map(c => `
-                        <li class="config-item">
-                            <span class="name">${escapeHtml(c)}</span>
-                            <button class="btn btn-danger" onclick="deactivateConfig('${escapeHtml(c).replace(/'/g, "\\'")}')">Deactivate</button>
-                        </li>
-                    `).join('');
-                } else if (!data.main_config && !data.error) {
-                    activeHtml += '<li class="config-item">No configurations in config.d/</li>';
-                }
-
-                activeEl.innerHTML = activeHtml || '<li class="config-item">No configurations found</li>';
-
-                if (availableConfigs.length === 0) {
-                    availEl.innerHTML = '<li class="config-item">No configurations in available.d/</li>';
-                } else {
-                    availEl.innerHTML = availableConfigs.map(c => `
-                        <li class="config-item">
-                            <span class="name">${escapeHtml(c)}</span>
-                            <button class="btn btn-success" onclick="activateConfig('${escapeHtml(c).replace(/'/g, "\\'")}')">Activate</button>
-                        </li>
-                    `).join('');
-                }
-            } catch (e) {
-                console.error('Error fetching configs:', e);
-                document.getElementById('active-configs').innerHTML = '<li class="config-item" style="color: var(--warning);">Error loading configs</li>';
-                document.getElementById('available-configs').innerHTML = '<li class="config-item" style="color: var(--warning);">Error loading configs</li>';
-            }
-        }
-
-        async function refreshHardware() {
-            const el = document.getElementById('hardware-list');
-            try {
-                const resp = await fetch('/api/hardware');
-                if (!resp.ok) {
-                    el.innerHTML = '<div class="hardware-item"><span class="badge warning">ERROR</span><span>Failed to fetch hardware status</span></div>';
-                    return;
-                }
-                const data = await resp.json();
-
-                // Handle error response
-                if (data.error) {
-                    el.innerHTML = `<div class="hardware-item"><span class="badge warning">ERROR</span><span>${escapeHtml(data.error)}</span></div>`;
-                    return;
-                }
-
-                const devices = Array.isArray(data.devices) ? data.devices : [];
-
-                if (devices.length === 0) {
-                    el.innerHTML = '<div class="hardware-item"><span class="badge info">INFO</span><span>No hardware devices detected (check if meshtasticd is running)</span></div>';
-                    return;
-                }
-
-                el.innerHTML = devices.map(d => `
-                    <div class="hardware-item">
-                        <span class="badge ${escapeHtml(d.type || 'info').toLowerCase()}">${escapeHtml(d.type || 'DEVICE')}</span>
-                        <span><strong>${escapeHtml(d.device || 'Unknown')}</strong> - ${escapeHtml(d.description || '')}</span>
-                    </div>
-                `).join('');
-            } catch (e) {
-                console.error('Error fetching hardware:', e);
-                el.innerHTML = '<div class="hardware-item"><span class="badge danger">ERROR</span><span>Network error fetching hardware</span></div>';
-            }
-        }
-
-        async function refreshRadio(forceRefresh = false) {
-            const el = document.getElementById('radio-info');
-            const btn = document.querySelector('[onclick*="refreshRadio"]');
-
-            // Show loading state
-            el.innerHTML = '<div class="item" style="grid-column: span 2;"><em>Loading radio info... (may take up to 30s)</em></div>';
-            if (btn) btn.disabled = true;
-
-            try {
-                const url = forceRefresh ? '/api/radio?refresh=1' : '/api/radio';
-                const resp = await fetch(url);
-                const data = await resp.json();
-
-                if (data.error) {
-                    el.innerHTML = `<div class="item" style="grid-column: span 2; color: var(--warning);">${escapeHtml(data.error)}</div>`;
-                } else {
-                    el.innerHTML = Object.entries(data).map(([k, v]) => `
-                        <div class="item">
-                            <div class="label">${escapeHtml(k.replace('_', ' ').toUpperCase())}</div>
-                            <div>${escapeHtml(v)}</div>
-                        </div>
-                    `).join('');
-                }
-            } catch (e) {
-                console.error('Error fetching radio:', e);
-                el.innerHTML = '<div class="item" style="grid-column: span 2; color: var(--danger);">Network error fetching radio info</div>';
-            } finally {
-                if (btn) btn.disabled = false;
-            }
-        }
-
-        async function refreshProcesses() {
-            try {
-                const resp = await fetch('/api/processes');
-                const data = await resp.json();
-                if (data.error) {
-                    document.getElementById('processes').textContent = 'Error: ' + data.error;
-                } else if (data.processes && Array.isArray(data.processes)) {
-                    document.getElementById('processes').textContent = data.processes.join('\\n');
-                } else {
-                    document.getElementById('processes').textContent = 'No process data available';
-                }
-            } catch (e) {
-                console.error('Error fetching processes:', e);
-                document.getElementById('processes').textContent = 'Error fetching processes';
-            }
-        }
-
-        // ========================================
-        // Gateway functions
-        // ========================================
-        async function refreshGateway() {
-            const statusEl = document.getElementById('gateway-status');
-            const meshEl = document.getElementById('gateway-mesh-status');
-            const rnsEl = document.getElementById('gateway-rns-status');
-            const nodesEl = document.getElementById('gateway-nodes');
-
-            try {
-                const resp = await fetch('/api/gateway/status');
-                if (!resp.ok) {
-                    throw new Error(`HTTP ${resp.status}`);
-                }
-                const data = await resp.json();
-
-                // Check for API error
-                if (data.error) {
-                    statusEl.innerHTML = `<span style="color: var(--danger);">Error</span>`;
-                    meshEl.innerHTML = '<span style="color: var(--text-muted);">--</span>';
-                    rnsEl.innerHTML = '<span style="color: var(--text-muted);">--</span>';
-                    console.error('Gateway API error:', data.error);
-                    return;
-                }
-
-                // Gateway status
-                if (data.running) {
-                    statusEl.innerHTML = '<span style="color: var(--accent);">Running</span>';
-                } else if (data.enabled) {
-                    statusEl.innerHTML = '<span style="color: var(--warning);">Stopped</span>';
-                } else {
-                    statusEl.innerHTML = '<span style="color: var(--text-muted);">Disabled</span>';
-                }
-
-                // Service status - check if services object exists
-                if (data.services) {
-                    meshEl.innerHTML = data.services.meshtasticd
-                        ? '<span style="color: var(--accent);">Online</span>'
-                        : '<span style="color: var(--danger);">Offline</span>';
-
-                    rnsEl.innerHTML = data.services.rnsd
-                        ? '<span style="color: var(--accent);">Online</span>'
-                        : '<span style="color: var(--danger);">Offline</span>';
-                } else {
-                    meshEl.innerHTML = '<span style="color: var(--text-muted);">Unknown</span>';
-                    rnsEl.innerHTML = '<span style="color: var(--text-muted);">Unknown</span>';
-                }
-
-                // Node count
-                nodesEl.textContent = data.stats?.total_nodes || 0;
-
-                // Update statistics
-                document.getElementById('stat-mesh-to-rns').textContent = data.stats?.mesh_to_rns || 0;
-                document.getElementById('stat-rns-to-mesh').textContent = data.stats?.rns_to_mesh || 0;
-                document.getElementById('stat-mesh-nodes').textContent = data.stats?.meshtastic_nodes || 0;
-                document.getElementById('stat-rns-nodes').textContent = data.stats?.rns_nodes || 0;
-                document.getElementById('stat-errors').textContent = data.stats?.errors || 0;
-                document.getElementById('stat-last-activity').textContent = data.stats?.last_activity || '--';
-
-                // Update button states
-                document.getElementById('gateway-start-btn').disabled = data.running;
-                document.getElementById('gateway-stop-btn').disabled = !data.running;
-
-            } catch (e) {
-                console.error('Error fetching gateway status:', e);
-                // Show error state in UI instead of leaving stale data
-                statusEl.innerHTML = '<span style="color: var(--danger);">Error</span>';
-                meshEl.innerHTML = '<span style="color: var(--text-muted);">--</span>';
-                rnsEl.innerHTML = '<span style="color: var(--text-muted);">--</span>';
-            }
-        }
-
-        async function startGateway() {
-            try {
-                document.getElementById('gateway-start-btn').disabled = true;
-                const resp = await fetch('/api/gateway/start', { method: 'POST' });
-                const data = await resp.json();
-                if (data.success) {
-                    alert(data.message || 'Gateway started');
-                } else {
-                    alert('Error: ' + (data.error || 'Failed to start gateway'));
-                }
-                refreshGateway();
-            } catch (e) {
-                alert('Error starting gateway: ' + e);
-                document.getElementById('gateway-start-btn').disabled = false;
-            }
-        }
-
-        async function stopGateway() {
-            try {
-                document.getElementById('gateway-stop-btn').disabled = true;
-                const resp = await fetch('/api/gateway/stop', { method: 'POST' });
-                const data = await resp.json();
-                if (data.success) {
-                    alert(data.message || 'Gateway stopped');
-                } else {
-                    alert('Error: ' + (data.error || 'Failed to stop gateway'));
-                }
-                refreshGateway();
-            } catch (e) {
-                alert('Error stopping gateway: ' + e);
-                document.getElementById('gateway-stop-btn').disabled = false;
-            }
-        }
-
-        async function runGatewayDiagnostic() {
-            const diagEl = document.getElementById('gateway-diagnostic');
-            diagEl.innerHTML = '<em>Running diagnostic checks...</em>';
-
-            try {
-                const resp = await fetch('/api/gateway/diagnostic', { method: 'POST' });
-                const data = await resp.json();
-
-                let html = '';
-                for (const result of data.results || []) {
-                    const color = result.status === 'PASS' ? 'var(--accent)' :
-                                  result.status === 'FAIL' ? 'var(--danger)' : 'var(--warning)';
-                    html += `<div style="margin: 5px 0;"><span style="color: ${color}; font-weight: bold;">[${result.status}]</span> ${escapeHtml(result.check)}: ${escapeHtml(result.message)}</div>`;
-                }
-
-                if (data.summary) {
-                    html += `<hr style="border-color: #444; margin: 10px 0;">`;
-                    html += `<div><strong>Summary:</strong> ${data.summary.passed} passed, ${data.summary.failed} failed, ${data.summary.warnings} warnings</div>`;
-                    if (data.summary.ready) {
-                        html += `<div style="color: var(--accent); margin-top: 5px;">✓ Gateway ready to start</div>`;
-                    } else {
-                        html += `<div style="color: var(--danger); margin-top: 5px;">✗ Fix issues before starting gateway</div>`;
-                    }
-                }
-
-                diagEl.innerHTML = html;
-            } catch (e) {
-                diagEl.innerHTML = `<span style="color: var(--danger);">Error running diagnostic: ${e}</span>`;
-            }
-        }
-
-        async function testGatewayConnections() {
-            const diagEl = document.getElementById('gateway-diagnostic');
-            diagEl.innerHTML = '<em>Testing connections...</em>';
-
-            try {
-                const resp = await fetch('/api/gateway/test', { method: 'POST' });
-                const data = await resp.json();
-
-                let html = '<div style="font-weight: bold; margin-bottom: 10px;">Connection Test Results:</div>';
-
-                // Meshtastic
-                const meshColor = data.meshtastic?.success ? 'var(--accent)' : 'var(--danger)';
-                html += `<div style="margin: 5px 0;"><span style="color: ${meshColor};">[${data.meshtastic?.success ? 'OK' : 'FAIL'}]</span> Meshtastic: ${escapeHtml(data.meshtastic?.message || 'Unknown')}</div>`;
-
-                // RNS
-                const rnsColor = data.rns?.success ? 'var(--accent)' : 'var(--danger)';
-                html += `<div style="margin: 5px 0;"><span style="color: ${rnsColor};">[${data.rns?.success ? 'OK' : 'FAIL'}]</span> RNS: ${escapeHtml(data.rns?.message || 'Unknown')}</div>`;
-
-                diagEl.innerHTML = html;
-            } catch (e) {
-                diagEl.innerHTML = `<span style="color: var(--danger);">Error testing connections: ${e}</span>`;
-            }
-        }
-
-        async function refreshGatewayNodes() {
-            const listEl = document.getElementById('gateway-node-list');
-            listEl.innerHTML = '<em>Loading nodes...</em>';
-
-            try {
-                const resp = await fetch('/api/gateway/nodes');
-                const data = await resp.json();
-
-                if (!data.nodes || data.nodes.length === 0) {
-                    listEl.innerHTML = '<div style="color: var(--text-muted); padding: 10px;">No nodes tracked yet. Start the gateway to begin tracking.</div>';
-                    return;
-                }
-
-                let html = `<table style="width: 100%; border-collapse: collapse;">
-                    <tr style="border-bottom: 1px solid #444;">
-                        <th style="padding: 8px; text-align: left;">Network</th>
-                        <th style="padding: 8px; text-align: left;">ID</th>
-                        <th style="padding: 8px; text-align: left;">Name</th>
-                        <th style="padding: 8px; text-align: left;">Last Seen</th>
-                    </tr>`;
-
-                for (const node of data.nodes) {
-                    const networkBadge = node.network === 'meshtastic'
-                        ? '<span style="background: #2196F3; color: white; padding: 2px 6px; border-radius: 3px; font-size: 0.8em;">MESH</span>'
-                        : '<span style="background: #9C27B0; color: white; padding: 2px 6px; border-radius: 3px; font-size: 0.8em;">RNS</span>';
-
-                    html += `<tr style="border-bottom: 1px solid #333;">
-                        <td style="padding: 8px;">${networkBadge}</td>
-                        <td style="padding: 8px; font-family: monospace;">${escapeHtml(node.id || '')}</td>
-                        <td style="padding: 8px;">${escapeHtml(node.name || '--')}</td>
-                        <td style="padding: 8px;">${escapeHtml(node.last_seen || '--')}</td>
-                    </tr>`;
-                }
-
-                html += '</table>';
-                html += `<div style="color: var(--text-muted); padding: 10px;">Total: ${data.total} nodes (${data.meshtastic_count} Meshtastic, ${data.rns_count} RNS)</div>`;
-
-                listEl.innerHTML = html;
-            } catch (e) {
-                listEl.innerHTML = `<span style="color: var(--danger);">Error loading nodes: ${e}</span>`;
-            }
-        }
-
-        async function refreshVersions() {
-            const listEl = document.getElementById('version-list');
-            const statusEl = document.getElementById('version-status');
-
-            listEl.innerHTML = '<em>Checking versions... (this may take a moment)</em>';
-            statusEl.textContent = '';
-
-            try {
-                const resp = await fetch('/api/versions');
-                const data = await resp.json();
-
-                if (data.error) {
-                    listEl.innerHTML = `<div style="color: var(--warning);">${escapeHtml(data.error)}</div>`;
-                    return;
-                }
-
-                // Defensive check for components array
-                if (!data.components || !Array.isArray(data.components)) {
-                    listEl.innerHTML = '<div style="color: var(--warning);">No version data available</div>';
-                    return;
-                }
-
-                let html = '<table style="width: 100%; border-collapse: collapse;">';
-                html += `
-                    <tr style="border-bottom: 1px solid #444;">
-                        <th style="padding: 10px; text-align: left;">Component</th>
-                        <th style="padding: 10px; text-align: left;">Installed</th>
-                        <th style="padding: 10px; text-align: left;">Latest</th>
-                        <th style="padding: 10px; text-align: left;">Status</th>
-                    </tr>
-                `;
-
-                data.components.forEach(c => {
-                    let statusBadge = '';
-                    if (c.installed === 'Not installed') {
-                        statusBadge = '<span style="color: var(--text-muted);">Not installed</span>';
-                    } else if (c.update_available) {
-                        statusBadge = '<span style="background: var(--warning); color: black; padding: 2px 8px; border-radius: 3px; font-size: 12px;">Update Available</span>';
-                    } else if (c.latest !== 'Unknown') {
-                        statusBadge = '<span style="color: var(--accent);">Up to date</span>';
-                    } else {
-                        statusBadge = '<span style="color: var(--text-muted);">--</span>';
-                    }
-
-                    html += `
-                        <tr style="border-bottom: 1px solid #333;">
-                            <td style="padding: 10px;"><strong>${escapeHtml(c.name)}</strong></td>
-                            <td style="padding: 10px; font-family: monospace;">${escapeHtml(c.installed)}</td>
-                            <td style="padding: 10px; font-family: monospace;">${escapeHtml(c.latest)}</td>
-                            <td style="padding: 10px;">${statusBadge}</td>
-                        </tr>
-                    `;
-                });
-
-                html += '</table>';
-                listEl.innerHTML = html;
-
-                // Status message
-                if (data.updates_available > 0) {
-                    statusEl.innerHTML = `<span style="color: var(--warning);">${data.updates_available} update(s) available</span>`;
-                } else {
-                    statusEl.innerHTML = '<span style="color: var(--accent);">All components up to date</span>';
-                }
-
-                statusEl.innerHTML += ` <span style="color: var(--text-muted);">| Last checked: ${new Date(data.checked_at).toLocaleTimeString()}</span>`;
-
-            } catch (e) {
-                console.error('Error fetching versions:', e);
-                listEl.innerHTML = `<div style="color: var(--danger);">Error: ${escapeHtml(e.message)}</div>`;
-            }
-        }
-
-        async function serviceAction(action) {
-            if (!confirm(`Are you sure you want to ${action} the service?`)) return;
-
-            try {
-                const resp = await fetch(`/api/service/${action}`, { method: 'POST' });
-                const data = await resp.json();
-                alert(data.success ? data.message : data.error);
-                fetchStatus();
-            } catch (e) {
-                alert('Error: ' + e);
-            }
-        }
-
-        async function activateConfig(name) {
-            try {
-                const resp = await fetch('/api/config/activate', {
-                    method: 'POST',
-                    headers: { 'Content-Type': 'application/json' },
-                    body: JSON.stringify({ config: name })
-                });
-                const data = await resp.json();
-                alert(data.success ? data.message : data.error);
-                fetchConfigs();
-            } catch (e) {
-                alert('Error: ' + e);
-            }
-        }
-
-        async function deactivateConfig(name) {
-            if (!confirm(`Deactivate ${name}?`)) return;
-            try {
-                const resp = await fetch('/api/config/deactivate', {
-                    method: 'POST',
-                    headers: { 'Content-Type': 'application/json' },
-                    body: JSON.stringify({ config: name })
-                });
-                const data = await resp.json();
-                alert(data.success ? data.message : data.error);
-                fetchConfigs();
-            } catch (e) {
-                alert('Error: ' + e);
-            }
-        }
-
-        async function refreshNodes() {
-            const el = document.getElementById('nodes-list');
-            const rawEl = document.getElementById('nodes-raw');
-            const rawContent = document.getElementById('nodes-raw-content');
-
-            el.innerHTML = '<em>Loading nodes... (may take up to 30s)</em>';
-
-            try {
-                // Use full nodes endpoint for detailed information
-                const resp = await fetch('/api/nodes/full');
-                const data = await resp.json();
-
-                if (data.error) {
-                    el.innerHTML = `<div style="color: var(--warning);">${data.error}</div>`;
-                } else if (data.nodes && Array.isArray(data.nodes) && data.nodes.length > 0) {
-                    el.innerHTML = `
-                        <table style="width: 100%; border-collapse: collapse;">
-                            <tr style="border-bottom: 1px solid #444;">
-                                <th style="padding: 8px; text-align: left;">Node ID</th>
-                                <th style="padding: 8px; text-align: left;">Name</th>
-                                <th style="padding: 8px; text-align: left;">Hardware</th>
-                                <th style="padding: 8px; text-align: left;">Battery</th>
-                                <th style="padding: 8px; text-align: left;">SNR</th>
-                                <th style="padding: 8px; text-align: left;">Hops</th>
-                                <th style="padding: 8px; text-align: left;">Last Heard</th>
-                            </tr>
-                            ${data.nodes.map(n => `
-                                <tr style="border-bottom: 1px solid #333;">
-                                    <td style="padding: 8px; font-family: monospace;">${escapeHtml(n.id || '')}</td>
-                                    <td style="padding: 8px;">${escapeHtml(n.name || n.short || '--')}</td>
-                                    <td style="padding: 8px;">${escapeHtml(n.hardware || '--')}</td>
-                                    <td style="padding: 8px;">${n.battery ? n.battery + '%' : '--'}</td>
-                                    <td style="padding: 8px;">${n.snr != null ? n.snr + ' dB' : '--'}</td>
-                                    <td style="padding: 8px;">${n.hops != null ? n.hops : '--'}</td>
-                                    <td style="padding: 8px;">${escapeHtml(n.last_heard_ago || '--')}</td>
-                                </tr>
-                            `).join('')}
-                        </table>
-                        <p style="margin-top: 10px; color: var(--text-muted);">${data.total_nodes || data.nodes.length} node(s) found${data.nodes_with_position ? `, ${data.nodes_with_position} with GPS` : ''}</p>
-                    `;
-                } else {
-                    el.innerHTML = '<div style="color: var(--text-muted);">No nodes found. Make sure meshtasticd is running and connected to a radio.</div>';
-                }
-
-                // Show raw output if available
-                if (data.raw) {
-                    rawEl.style.display = 'block';
-                    rawContent.textContent = data.raw;
-                }
-            } catch (e) {
-                console.error('Error fetching nodes:', e);
-                el.innerHTML = '<div style="color: var(--danger);">Network error fetching nodes</div>';
-            }
-        }
-
-        async function sendMessage() {
-            const text = document.getElementById('message-text').value;
-            const dest = document.getElementById('message-dest').value;
-            const statusEl = document.getElementById('message-status');
-
-            if (!text.trim()) {
-                statusEl.innerHTML = '<span style="color: var(--warning);">Please enter a message</span>';
-                return;
-            }
-
-            statusEl.innerHTML = '<em>Sending...</em>';
-
-            try {
-                const resp = await fetch('/api/message', {
-                    method: 'POST',
-                    headers: { 'Content-Type': 'application/json' },
-                    body: JSON.stringify({ text: text, destination: dest || null })
-                });
-                const data = await resp.json();
-
-                if (data.success) {
-                    statusEl.innerHTML = '<span style="color: var(--accent);">Message sent successfully!</span>';
-                    document.getElementById('message-text').value = '';
-                } else {
-                    statusEl.innerHTML = `<span style="color: var(--danger);">${data.error}</span>`;
-                }
-            } catch (e) {
-                console.error('Error sending message:', e);
-                statusEl.innerHTML = '<span style="color: var(--danger);">Network error sending message</span>';
-            }
-        }
-
-        // Initial load
-        fetchStatus();
-        fetchLogs();
-        fetchConfigs();
-        refreshHardware();
-        refreshRadio();
-        refreshNodes();
-        refreshProcesses();
-        refreshGateway();
-
-        // Auto-refresh status every 5 seconds
-        setInterval(fetchStatus, 5000);
-        setInterval(refreshGateway, 10000);  // Refresh gateway status every 10s
-
-        // ========================================
-        // Map functionality
-        // ========================================
-        let meshMap = null;
-        let mapMarkers = [];
-        let mapInitialized = false;
-
-        function initMap() {
-            if (mapInitialized) return;
-
-            const mapContainer = document.getElementById('mesh-map');
-            if (!mapContainer) return;
-
-            // Check if Leaflet is loaded
-            if (typeof L === 'undefined') {
-                console.error('Leaflet not loaded yet');
-                mapContainer.innerHTML = '<div style="color: var(--warning); padding: 20px;">Map library loading... Refresh the page if this persists.</div>';
-                return;
-            }
-
-            // Initialize map centered on a default location
-            meshMap = L.map('mesh-map').setView([20, 0], 2);
-
-            // Add OpenStreetMap tiles with dark theme
-            L.tileLayer('https://{s}.basemaps.cartocdn.com/dark_all/{z}/{x}/{y}{r}.png', {
-                attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/">CARTO</a>',
-                subdomains: 'abcd',
-                maxZoom: 19
-            }).addTo(meshMap);
-
-            mapInitialized = true;
-            refreshMap();
-        }
-
-        function getNodeColor(node) {
-            if (node.is_me) return '#4CAF50';  // Green for my node
-
-            if (!node.last_heard) return '#607D8B';  // Gray for unknown
-
-            const lastHeard = new Date(node.last_heard);
-            const now = new Date();
-            const hoursSince = (now - lastHeard) / (1000 * 60 * 60);
-
-            if (hoursSince < 1) return '#2196F3';   // Blue for online
-            if (hoursSince < 24) return '#ff9800';  // Orange for stale
-            return '#607D8B';                        // Gray for offline
-        }
-
-        function createNodeIcon(color, isMe) {
-            const size = isMe ? 16 : 12;
-            return L.divIcon({
-                className: 'custom-marker',
-                html: `<div style="
-                    width: ${size}px;
-                    height: ${size}px;
-                    background: ${color};
-                    border: 2px solid white;
-                    border-radius: 50%;
-                    box-shadow: 0 2px 5px rgba(0,0,0,0.5);
-                "></div>`,
-                iconSize: [size, size],
-                iconAnchor: [size/2, size/2],
-                popupAnchor: [0, -size/2]
-            });
-        }
-
-        async function refreshMap() {
-            const countEl = document.getElementById('map-node-count');
-            const listEl = document.getElementById('map-node-list');
-
-            // Check if map is initialized
-            if (!meshMap || typeof L === 'undefined') {
-                countEl.textContent = 'Map not initialized';
-                listEl.innerHTML = '<div style="color: var(--warning);">Click the Map tab to initialize the map first.</div>';
-                return;
-            }
-
-            countEl.textContent = 'Loading...';
-
-            try {
-                // Add timeout to prevent infinite loading
-                const controller = new AbortController();
-                const timeoutId = setTimeout(() => controller.abort(), 30000);
-
-                const resp = await fetch('/api/nodes/full', { signal: controller.signal });
-                clearTimeout(timeoutId);
-                const data = await resp.json();
-
-                if (data.error) {
-                    countEl.textContent = data.error;
-                    listEl.innerHTML = `<div style="color: var(--warning);">${data.error}</div>`;
-                    return;
-                }
-
-                // Ensure nodes is an array
-                if (!data.nodes || !Array.isArray(data.nodes)) {
-                    countEl.textContent = 'No nodes available';
-                    listEl.innerHTML = '<div style="color: var(--text-muted);">No node data available. Connect to meshtasticd first.</div>';
-                    return;
-                }
-
-                // Clear existing markers
-                mapMarkers.forEach(m => meshMap.removeLayer(m));
-                mapMarkers = [];
-
-                // Filter nodes with positions
-                const nodesWithPos = data.nodes.filter(n => n.position && n.position.latitude && n.position.longitude);
-
-                countEl.textContent = `${nodesWithPos.length} of ${data.total_nodes} nodes with position`;
-
-                if (nodesWithPos.length === 0) {
-                    listEl.innerHTML = '<div style="color: var(--text-muted);">No nodes with GPS position found. Nodes need to have GPS or fixed position set.</div>';
-                    return;
-                }
-
-                // Add markers
-                const bounds = [];
-                nodesWithPos.forEach(node => {
-                    const lat = node.position.latitude;
-                    const lng = node.position.longitude;
-                    const color = getNodeColor(node);
-                    const icon = createNodeIcon(color, node.is_me);
-
-                    const marker = L.marker([lat, lng], { icon: icon }).addTo(meshMap);
-
-                    // Build popup content
-                    let popupContent = `
-                        <div style="min-width: 200px; font-family: sans-serif;">
-                            <h3 style="margin: 0 0 8px 0; color: #333;">${node.name}</h3>
-                            <table style="font-size: 12px; color: #666;">
-                                <tr><td style="padding-right: 10px;"><strong>ID:</strong></td><td><code>${node.id}</code></td></tr>
-                                <tr><td><strong>Short:</strong></td><td>${node.short || '--'}</td></tr>
-                                <tr><td><strong>Hardware:</strong></td><td>${node.hardware || '--'}</td></tr>
-                    `;
-
-                    if (node.battery) {
-                        popupContent += `<tr><td><strong>Battery:</strong></td><td>${node.battery}%</td></tr>`;
-                    }
-                    if (node.last_heard_ago) {
-                        popupContent += `<tr><td><strong>Last Heard:</strong></td><td>${node.last_heard_ago}</td></tr>`;
-                    }
-                    if (node.snr !== null && node.snr !== undefined) {
-                        popupContent += `<tr><td><strong>SNR:</strong></td><td>${node.snr} dB</td></tr>`;
-                    }
-                    if (node.hops !== null && node.hops !== undefined) {
-                        popupContent += `<tr><td><strong>Hops:</strong></td><td>${node.hops}</td></tr>`;
-                    }
-                    if (node.position.altitude) {
-                        popupContent += `<tr><td><strong>Altitude:</strong></td><td>${node.position.altitude}m</td></tr>`;
-                    }
-
-                    popupContent += `
-                                <tr><td><strong>Position:</strong></td><td>${lat.toFixed(5)}, ${lng.toFixed(5)}</td></tr>
-                            </table>
-                        </div>
-                    `;
-
-                    marker.bindPopup(popupContent);
-                    mapMarkers.push(marker);
-                    bounds.push([lat, lng]);
-                });
-
-                // Fit map to bounds if we have markers
-                if (bounds.length > 0) {
-                    if (bounds.length === 1) {
-                        meshMap.setView(bounds[0], 14);
-                    } else {
-                        meshMap.fitBounds(bounds, { padding: [50, 50] });
-                    }
-                }
-
-                // Update node list
-                listEl.innerHTML = nodesWithPos.map(n => `
-                    <div class="hardware-item" style="cursor: pointer;" onclick="focusNode(${n.position.latitude}, ${n.position.longitude})">
-                        <span class="badge" style="background: ${getNodeColor(n)};">${n.is_me ? 'ME' : n.short || '??'}</span>
-                        <span>
-                            <strong>${n.name}</strong>
-                            <span style="color: var(--text-muted); margin-left: 10px;">
-                                ${n.position.latitude.toFixed(4)}, ${n.position.longitude.toFixed(4)}
-                                ${n.last_heard_ago ? ' - ' + n.last_heard_ago : ''}
-                            </span>
-                        </span>
-                    </div>
-                `).join('');
-
-            } catch (e) {
-                console.error('Error refreshing map:', e);
-                if (e.name === 'AbortError') {
-                    countEl.textContent = 'Timeout - try again';
-                    listEl.innerHTML = '<div style="color: var(--warning);">Connection timed out. Click Refresh Map to try again.</div>';
-                } else {
-                    countEl.textContent = 'Error loading map data';
-                    listEl.innerHTML = `<div style="color: var(--danger);">Error: ${e.message}</div>`;
-                }
-            }
-        }
-
-        function focusNode(lat, lng) {
-            if (meshMap) {
-                meshMap.setView([lat, lng], 16);
-            }
-        }
-
-        // Initialize map when Map tab is clicked
-        document.querySelector('.tab[data-tab="map"]').addEventListener('click', () => {
-            setTimeout(() => {
-                initMap();
-                if (meshMap) {
-                    meshMap.invalidateSize();
-                }
-            }, 100);
-        });
-    </script>
-</body>
-</html>
-'''
 
 
 # ============================================================================
@@ -2621,8 +1034,8 @@ def favicon():
 @login_required
 def index():
     from __version__ import __version__
-    return render_template_string(
-        MAIN_TEMPLATE,
+    return render_template(
+        'main.html',
         version=__version__,
         auth_enabled=CONFIG['auth_enabled']
     )
